pyproject.toml
[project]
name = "tou-document-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "werkzeug>=3.1.3",
]
